// src/types/gltf.d.ts
export type GlTF = GlTFProperty & {
  /**
   * Names of glTF extensions used in this asset.
   *
   * @minItems 1
   */
  extensionsUsed?: [string, ...string[]];
  /**
   * Names of glTF extensions required to properly load this asset.
   *
   * @minItems 1
   */
  extensionsRequired?: [string, ...string[]];
  /**
   * An array of accessors.
   *
   * @minItems 1
   */
  accessors?: [Accessor, ...Accessor[]];
  /**
   * An array of keyframe animations.
   *
   * @minItems 1
   */
  animations?: [Animation, ...Animation[]];
  /**
   * Metadata about the glTF asset.
   */
  asset: Asset;
  /**
   * An array of buffers.
   *
   * @minItems 1
   */
  buffers?: [Buffer, ...Buffer[]];
  /**
   * An array of bufferViews.
   *
   * @minItems 1
   */
  bufferViews?: [BufferView, ...BufferView[]];
  /**
   * An array of cameras.
   *
   * @minItems 1
   */
  cameras?: [Camera, ...Camera[]];
  /**
   * An array of images.
   *
   * @minItems 1
   */
  images?: [Image, ...Image[]];
  /**
   * An array of materials.
   *
   * @minItems 1
   */
  materials?: [Material, ...Material[]];
  /**
   * An array of meshes.
   *
   * @minItems 1
   */
  meshes?: [Mesh, ...Mesh[]];
  /**
   * An array of nodes.
   *
   * @minItems 1
   */
  nodes?: [Node, ...Node[]];
  /**
   * An array of samplers.
   *
   * @minItems 1
   */
  samplers?: [Sampler, ...Sampler[]];
  /**
   * The index of the default scene.
   */
  scene?: GlTFId;
  /**
   * An array of scenes.
   *
   * @minItems 1
   */
  scenes?: [Scene, ...Scene[]];
  /**
   * An array of skins.
   *
   * @minItems 1
   */
  skins?: [Skin, ...Skin[]];
  /**
   * An array of textures.
   *
   * @minItems 1
   */
  textures?: [Texture, ...Texture[]];
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
// ...
export type GlTFId = number;
// ...
export type Animation = GlTFChildOfRootProperty & {
  /**
   * An array of animation channels. An animation channel combines an animation sampler with a target property being animated. Different channels of the same animation **MUST NOT** have the same targets.
   *
   * @minItems 1
   */
  channels: [AnimationChannel, ...AnimationChannel[]];
  /**
   * An array of animation samplers. An animation sampler combines timestamps with a sequence of output values and defines an interpolation algorithm.
   *
   * @minItems 1
   */
  samplers: [AnimationSampler, ...AnimationSampler[]];
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & {
  /**
   * An array of animation channels. An animation channel combines an animation sampler with a target property being animated. Different channels of the same animation **MUST NOT** have the same targets.
   *
   * @minItems 1
   */
  channels: [AnimationChannel, ...AnimationChannel[]];
  /**
   * An array of animation samplers. An animation sampler combines timestamps with a sequence of output values and defines an interpolation algorithm.
   *
   * @minItems 1
   */
  samplers: [AnimationSampler, ...AnimationSampler[]];
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
// ...
export type Asset = GlTFProperty & {
  /**
   * A copyright message suitable for display to credit the content creator.
   */
  copyright?: string;
  /**
   * Tool that generated this glTF model.  Useful for debugging.
   */
  generator?: string;
  /**
   * The glTF version in the form of `<major>.<minor>` that this asset targets.
   */
  version: string;
  /**
   * The minimum glTF version in the form of `<major>.<minor>` that this asset targets. This property **MUST NOT** be greater than the asset version.
   */
  minVersion?: string;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
/**
 * A buffer points to binary geometry, animation, or skins.
 */
export type Buffer = GlTFChildOfRootProperty & {
  /**
   * The URI (or IRI) of the buffer.
   */
  uri?: string;
  /**
   * The length of the buffer in bytes.
   */
  byteLength: number;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & {
  /**
   * The URI (or IRI) of the buffer.
   */
  uri?: string;
  /**
   * The length of the buffer in bytes.
   */
  byteLength: number;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
/**
 * A view into a buffer generally representing a subset of the buffer.
 */
export type BufferView = GlTFChildOfRootProperty & {
  /**
   * The index of the buffer.
   */
  buffer: GlTFId;
  /**
   * The offset into the buffer in bytes.
   */
  byteOffset?: number;
  /**
   * The length of the bufferView in bytes.
   */
  byteLength: number;
  /**
   * The stride, in bytes.
   */
  byteStride?: number;
  /**
   * The hint representing the intended GPU buffer type to use with this buffer view.
   */
  target?: 34962 | 34963 | number;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & {
  /**
   * The index of the buffer.
   */
  buffer: GlTFId;
  /**
   * The offset into the buffer in bytes.
   */
  byteOffset?: number;
  /**
   * The length of the bufferView in bytes.
   */
  byteLength: number;
  /**
   * The stride, in bytes.
   */
  byteStride?: number;
  /**
   * The hint representing the intended GPU buffer type to use with this buffer view.
   */
  target?: 34962 | 34963 | number;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
/**
 * A camera's projection.  A node **MAY** reference a camera to apply a transform to place the camera in the scene.
 */
export type Camera = GlTFChildOfRootProperty & {
  /**
   * An orthographic camera containing properties to create an orthographic projection matrix. This property **MUST NOT** be defined when `perspective` is defined.
   */
  orthographic?: CameraOrthographic;
  /**
   * A perspective camera containing properties to create a perspective projection matrix. This property **MUST NOT** be defined when `orthographic` is defined.
   */
  perspective?: CameraPerspective;
  /**
   * Specifies if the camera uses a perspective or orthographic projection.
   */
  type: "perspective" | "orthographic" | string;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & {
  /**
   * An orthographic camera containing properties to create an orthographic projection matrix. This property **MUST NOT** be defined when `perspective` is defined.
   */
  orthographic?: CameraOrthographic;
  /**
   * A perspective camera containing properties to create a perspective projection matrix. This property **MUST NOT** be defined when `orthographic` is defined.
   */
  perspective?: CameraPerspective;
  /**
   * Specifies if the camera uses a perspective or orthographic projection.
   */
  type: "perspective" | "orthographic" | string;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
// ...
export type Image = GlTFChildOfRootProperty & {
  /**
   * The URI (or IRI) of the image.
   */
  uri?: string;
  /**
   * The image's media type. This field **MUST** be defined when `bufferView` is defined.
   */
  mimeType?: "image/jpeg" | "image/png" | string;
  /**
   * The index of the bufferView that contains the image. This field **MUST NOT** be defined when `uri` is defined.
   */
  bufferView?: GlTFId;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & (
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      }
  ) & {
    /**
     * The URI (or IRI) of the image.
     */
    uri?: string;
    /**
     * The image's media type. This field **MUST** be defined when `bufferView` is defined.
     */
    mimeType?: "image/jpeg" | "image/png" | string;
    /**
     * The index of the bufferView that contains the image. This field **MUST NOT** be defined when `uri` is defined.
     */
    bufferView?: GlTFId;
    name?: unknown;
    extensions?: unknown;
    extras?: unknown;
    [k: string]: unknown;
  } & (
    | {
        [k: string]: unknown;
      }
    | {
        [k: string]: unknown;
      }
  );
/**
 * The material appearance of a primitive.
 */
export type Material = GlTFChildOfRootProperty & {
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  /**
   * A set of parameter values that are used to define the metallic-roughness material model from Physically Based Rendering (PBR) methodology. When undefined, all the default values of `pbrMetallicRoughness` **MUST** apply.
   */
  pbrMetallicRoughness?: MaterialPBRMetallicRoughness;
  /**
   * The tangent space normal texture.
   */
  normalTexture?: MaterialNormalTextureInfo;
  /**
   * The occlusion texture.
   */
  occlusionTexture?: MaterialOcclusionTextureInfo;
  /**
   * The emissive texture.
   */
  emissiveTexture?: TextureInfo;
  /**
   * The factors for the emissive color of the material.
   *
   * @minItems 3
   * @maxItems 3
   */
  emissiveFactor?: [number, number, number];
  /**
   * The alpha rendering mode of the material.
   */
  alphaMode?: "OPAQUE" | "MASK" | "BLEND" | string;
  /**
   * The alpha cutoff value of the material.
   */
  alphaCutoff?: number;
  /**
   * Specifies whether the material is double sided.
   */
  doubleSided?: boolean;
  [k: string]: unknown;
} & {
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  /**
   * A set of parameter values that are used to define the metallic-roughness material model from Physically Based Rendering (PBR) methodology. When undefined, all the default values of `pbrMetallicRoughness` **MUST** apply.
   */
  pbrMetallicRoughness?: MaterialPBRMetallicRoughness;
  /**
   * The tangent space normal texture.
   */
  normalTexture?: MaterialNormalTextureInfo;
  /**
   * The occlusion texture.
   */
  occlusionTexture?: MaterialOcclusionTextureInfo;
  /**
   * The emissive texture.
   */
  emissiveTexture?: TextureInfo;
  /**
   * The factors for the emissive color of the material.
   *
   * @minItems 3
   * @maxItems 3
   */
  emissiveFactor?: [number, number, number];
  /**
   * The alpha rendering mode of the material.
   */
  alphaMode?: "OPAQUE" | "MASK" | "BLEND" | string;
  /**
   * The alpha cutoff value of the material.
   */
  alphaCutoff?: number;
  /**
   * Specifies whether the material is double sided.
   */
  doubleSided?: boolean;
  [k: string]: unknown;
};
// ...
export type Mesh = GlTFChildOfRootProperty & {
  /**
   * An array of primitives, each defining geometry to be rendered.
   *
   * @minItems 1
   */
  primitives: [MeshPrimitive, ...MeshPrimitive[]];
  /**
   * Array of weights to be applied to the morph targets. The number of array elements **MUST** match the number of morph targets.
   *
   * @minItems 1
   */
  weights?: [number, ...number[]];
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & {
  /**
   * An array of primitives, each defining geometry to be rendered.
   *
   * @minItems 1
   */
  primitives: [MeshPrimitive, ...MeshPrimitive[]];
  /**
   * Array of weights to be applied to the morph targets. The number of array elements **MUST** match the number of morph targets.
   *
   * @minItems 1
   */
  weights?: [number, ...number[]];
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
// ...
export type Sampler = GlTFChildOfRootProperty & {
  /**
   * Magnification filter.
   */
  magFilter?: 9728 | 9729 | number;
  /**
   * Minification filter.
   */
  minFilter?: 9728 | 9729 | 9984 | 9985 | 9986 | 9987 | number;
  /**
   * S (U) wrapping mode.
   */
  wrapS?: 33071 | 33648 | 10497 | number;
  /**
   * T (V) wrapping mode.
   */
  wrapT?: 33071 | 33648 | 10497 | number;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
} & {
  /**
   * Magnification filter.
   */
  magFilter?: 9728 | 9729 | number;
  /**
   * Minification filter.
   */
  minFilter?: 9728 | 9729 | 9984 | 9985 | 9986 | 9987 | number;
  /**
   * S (U) wrapping mode.
   */
  wrapS?: 33071 | 33648 | 10497 | number;
  /**
   * T (V) wrapping mode.
   */
  wrapT?: 33071 | 33648 | 10497 | number;
  name?: unknown;
  extensions?: unknown;
  extras?: unknown;
  [k: string]: unknown;
};
// ...
export interface GlTFProperty {
  extensions?: Extension;
  extras?: Extras;
  [k: string]: unknown;
}